datascience-lab/cheats/requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
plotly
ipykernel
jupyter
streamlit
sqlalchemy
psycopg2-binary
python-dotenv